move relatedRestaurants logic from DetailPage component to Detail page
