+++ pages/detail/[id]/[name].tsx
@@ -1,8 +1,12 @@
 import React, { useState, useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
 import { NextPage, NextPageContext } from 'next';
 import Head from 'next/head';
 
 import DetailPage from '../../../components/DetailPage/DetailPage';
+
+import { InitialState } from '../../../store/reducer';
+import { clearRelatedRestaurants } from '../../../store/actions';
 
 import { RestaurantDetail } from '../../../helpers/types';
 
@@ -22,9 +26,20 @@ type CustomNextPageContext = NextPageContext & {
 const Detail: NextPage<DetailProps> = ({ data, id }) => {
   const [isLoading, setIsLoading] = useState(false);
 
+  const relatedRestaurants = useSelector(
+    (state: InitialState) => state.relatedRestaurants,
+  );
+
+  const dispatch = useDispatch();
+
   useEffect(() => {
     setIsLoading(false);
   }, [id]);
+
+  const handleClickRelatedRestaurant = () => {
+    setIsLoading(true);
+    dispatch(clearRelatedRestaurants());
+  };
 
   return (
     <>
@@ -34,7 +49,8 @@ const Detail: NextPage<DetailProps> = ({ data, id }) => {
       <DetailPage
         data={data}
         isLoading={isLoading}
-        onClickRelatedRestaurant={() => setIsLoading(true)}
+        relatedRestaurants={relatedRestaurants}
+        onClickRelatedRestaurant={handleClickRelatedRestaurant}
       />
     </>
   );
